+++ packages/analytics/src/Analytics.ts
@@ -25,8 +25,11 @@ import { PageViewTracker, EventTracker, SessionTracker } from './trackers';
 
 const logger = new Logger('AnalyticsClass');
 
-const dispatchAnalyticsEvent = (event, data) => {
-    Hub.dispatch('analytics', { event, data }, 'Analytics');
+const AMPLIFY_SYMBOL = ((typeof Symbol !== 'undefined' && typeof Symbol.for === 'function') ?
+    Symbol.for('amplify_default') : '@@amplify_default') as Symbol;
+
+const dispatchAnalyticsEvent = (event:string, data:any, message:string) => {
+    Hub.dispatch('analytics', { event, data, message }, 'Analytics', AMPLIFY_SYMBOL);
 };
 
 const trackers = {
@@ -93,7 +96,11 @@ export default class AnalyticsClass {
             this._config['autoSessionRecord'] = true;
         }
 
-        dispatchAnalyticsEvent('configured', null);
+        dispatchAnalyticsEvent(
+            'configured', 
+            null,
+            `The Analytics category has been configured successfully`
+        );
         logger.debug('current configuration', this._config);
 
         
@@ -167,13 +174,6 @@ export default class AnalyticsClass {
     public enable() {
         this._disabled = false;
     }
-
-    /**
-    * Receive a capsule from Hub
-    * @param {any} capsuak - The message from hub
-    */
-   public onHubCapsule(capsule: any): void {}
-
 
     /**
      * Record Session start