+++ packages/aws-amplify-react/src/Auth/SignIn.jsx
@@ -67,7 +67,7 @@ export default class SignIn extends AuthPiece {
         // avoid submitting the form
         event.preventDefault();
         
-        const { username, password } = this.inputs;
+        const { username='', password } = this.inputs;
         if (!Auth || typeof Auth.signIn !== 'function') {
             throw new Error('No Auth module found, please ensure @aws-amplify/auth is imported');
         }