+++ packages/core/src/ClientDevice/browser.ts
@@ -88,22 +88,3 @@ function browserType(userAgent) {
 
     return { type: '', version: '' };
 }
-
-if (JS.browserOrNode().isBrowser && typeof window.addEventListener === 'function') {
-    window.addEventListener('resize', function() {
-        Hub.dispatch('window', { event: 'resize', data: dimension() }, 'DeviceInfo');
-    });
-
-    window.addEventListener('scroll', function() {
-        const pos = { x: window.scrollX, y: window.scrollY };
-        Hub.dispatch('window', { event: 'scroll', data: pos }, 'DeviceInfo');
-    });
-
-    window.addEventListener('offline', function() {
-        Hub.dispatch('window', { event: 'offline' }, 'DeviceInfor');
-    });
-
-    window.addEventListener('online', function() {
-        Hub.dispatch('window', { event: 'online' }, 'DeviceInfor');
-    });
-}